Team-33-CodingGeeks-1/impl05/Team-33-CodingGeeks/requirements.txt

websockets
cryptography
orjson
pydantic
pyyaml
uvloop==0.19.0
aiosqlite
click
pytest
